blogapp/serializer.py
from .models import *
from rest_framework import serializers

class BlogSerializer(serializers.ModelSerializer):
    class Meta:
        model= Blog
        fields=['title','content','publishedDate','id']
        


