requirement.txt
asgiref==3.7.2
azure-core==1.12.0
azure-storage-blob==12.8.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
cycler==0.10.0
Django==3.2
django-filter==23.5
djangorestframework==3.11.0
dnspython==2.4.2
drf-yasg==1.20.0
future==0.18.2
idna==2.10
importlib-metadata==7.0.0
inflection==0.5.1
isodate==0.6.0
itypes==1.2.0
Jinja2==3.0.1
kiwisolver==1.3.1
Markdown==3.5.1
MarkupSafe==2.0.1
matplotlib==3.4.1
msrest==0.6.21
netifaces==0.10.9
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==21.0
Pillow==8.2.0
protobuf==3.14.0
pycparser==2.20
pymongo==4.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-igraph==0.9.6
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
schedule==1.1.0
six==1.15.0
sqlparse==0.4.2
texttable==1.6.3
tradingview-ta==3.3.0
treelib==1.6.1
typing_extensions==4.8.0
uritemplate==3.0.1
urllib3==1.26.4
zipp==3.17.0